Moved GcmHelper functionality to EasyGcm class and deprecated GcmHelper. Registration now only needs Context instead of an Activity
