+++ easygcm-lib/src/main/java/eu/inloop/easygcm/GcmServicesHandler.java
@@ -15,7 +15,7 @@ public class GcmServicesHandler {
      */
     protected void onPlayServicesUnavailable(Activity context, int errorCode, boolean recoverable) {
         GoogleApiAvailability.getInstance().getErrorDialog(context, errorCode, PLAY_SERVICES_RESOLUTION_REQUEST).show();
-        if (GcmHelper.sLoggingEnabled) {
+        if (EasyGcm.sLoggingEnabled) {
             GcmUtils.Logger.d("This device is not supported. Error code " + errorCode + " Recoverable - " + recoverable);
         }
     }
